Fixing student issues

+++ admin/view/test/test_center.php
@@ -33,7 +33,7 @@
                 <td> <?php echo $rows['pass_score']; ?></td>
                 <td>
                 <a href="?route=test/question&parameter=<?php echo $rows['test_id']; ?>">View</a>
-                  <a href="?route=test/test_center&method=deleteTest&parameter=<?php echo $rows['test_id']; ?>">Del</a>
+                  <a href="?route=test/test_center&method=deleteTest&parameter=<?php echo $rows['test_id']; ?>" onclick="return confirm('Are you sure you want to delete this item?');">Del</a>
                   <a href="?route=test/test_center&method=editTest&parameter=<?php echo $rows['test_id']; ?>">Edit</a>
                 </td>
           </tr>
